requirements.txt
PyYAML==6.0.1
torch==2.4.0
transformers==4.42.3
accelerate==0.31.0
requests==2.32.3
